Rename: CommunityCard 경로 변경
