+++ src/components/Community/CommunityCard.tsx
